moved rescale_value_array
