+++ ivadomed/maths.py
@@ -1,8 +1,28 @@
-import nibabel as nib
 import numpy as np
 import os
 import scipy
-import ivadomed.transforms as imed_transforms
+
+
+def rescale_values_array(arr, minv=0.0, maxv=1.0, dtype=np.float32):
+    """Rescale the values of numpy array `arr` to be from `minv` to `maxv`.
+
+    Args:
+        arr (ndarry): Array whose values will be rescaled.
+        minv (float): Minimum value of the output array.
+        maxv (float): Maximum value of the output array.
+        dtype (type): Cast array to this type before performing the rescaling.
+    """
+    if dtype is not None:
+        arr = arr.astype(dtype)
+
+    mina = np.min(arr)
+    maxa = np.max(arr)
+
+    if mina == maxa:
+        return arr * minv
+
+    norm = (arr - mina) / (maxa - mina)  # normalize the array first
+    return (norm * (maxv - minv)) + minv  # rescale by minv and maxv, which is the normalized array by default
 
 
 def gaussian_kernel(kernlen=10):
@@ -19,7 +39,7 @@ def gaussian_kernel(kernlen=10):
     x = np.linspace(-1, 1, kernlen + 1)
     kern1d = np.diff(scipy.stats.norm.cdf(x))
     kern2d = np.outer(kern1d, kern1d)
-    return imed_transforms.rescale_values_array(kern2d / kern2d.sum())
+    return rescale_values_array(kern2d / kern2d.sum())
 
 
 def heatmap_generation(image, kernel_size):
@@ -36,5 +56,5 @@ def heatmap_generation(image, kernel_size):
     """
     kernel = gaussian_kernel(kernel_size)
     map = scipy.signal.convolve(image, kernel, mode='same')
-    return imed_transforms.rescale_values_array(map)
+    return rescale_values_array(map)
